feat: better version track in library
