+++ Cargo.toml
@@ -5,14 +5,6 @@ license.workspace = true
 edition.workspace = true
 repository.workspace = true
 
-[lib]
-name = "pulsar"
-path = "src/lib.rs"
-
-[[bin]]
-name = "pulsar-exec"
-path = "src/main.rs"
-
 [dependencies]
 # Crates
 bpf-common = { workspace = true }
@@ -38,6 +30,11 @@ rust-ini = { workspace = true }
 serde = { workspace = true }
 semver = { workspace = true, features = ["serde"] }
 tokio = { workspace = true, features = ["full"] }
+clap-verbosity-flag = { workspace = true }
+
+[build-dependencies]
+anyhow = { workspace = true }
+vergen-gitcl = { workspace = true }
 
 [features]
 default = ["full", "tls-openssl"]
@@ -78,7 +75,7 @@ members = [
 ]
 
 [workspace.package]
-version = "0.9.0"
+version = "0.10.0-alpha.1"
 license = "Apache-2.0 WITH BPF probes exception under GPL-2.0"
 edition = "2024"
 repository = "https://github.com/exein-io/pulsar"
@@ -113,46 +110,45 @@ process-monitor = { path = "crates/modules/process-monitor", features = [
 rules-engine = { path = "crates/modules/rules-engine" }
 smtp-notifier = { path = "crates/modules/smtp-notifier" }
 # External
-anyhow = "1.0.75"
+anyhow = "1"
 aya = { version = "0.13.1", features = ["async_tokio"] }
 aya-ebpf-bindings = "0.1.0"
-aya-obj = "0.1.0"
-axum = { version = "0.8.1", features = ["ws"] }
-bytes = "1.5.0"
-chrono = { version = "0.4.31" }
-clap = { version = "4.4.11", features = ["derive"] }
-comfy-table = "7.1.0"
-dns-mock-server = "0.1.4"
+aya-obj = "0.2.1"
+axum = { version = "0.8.6", features = ["ws"] }
+bytes = "1.10.1"
+chrono = { version = "0.4.42" }
+clap = { version = "4.5.49", features = ["derive"] }
+comfy-table = "7.2.1"
+dns-mock-server = "0.1.5"
 dns-parser = "0.8.0"
-diesel = { version = "2.1", features = ["sqlite"] }
-env_logger = "0.10.1"
+diesel = { version = "2.3.2", features = ["sqlite"] }
+env_logger = "0.11.8"
 flate2 = "1.0"
 futures = "0.3.29"
 futures-util = "0.3.29"
-gethostname = "0.4.3"
-glob = "0.3.1"
+gethostname = "1.1.0"
+glob = "0.3.3"
 hex = "0.4.3"
-hickory-resolver = "0.24.1"
-http-body-util = "0.1.2"
-hyper = "1.5.1"
-hyper-util = "0.1.10"
+hickory-resolver = "0.25.2"
+http-body-util = "0.1.3"
+hyper = "1.7.0"
+hyper-util = "0.1.17"
 hyperlocal = "0.9.1"
-indicatif = "0.17"
-lalrpop = "0.20.0"
-lalrpop-util = { version = "0.20.0", features = ["lexer"] }
-lazy_static = "1.4"
-leon = "2.0.1"
-lettre = { version = "0.11.2", default-features = false, features = [
+indicatif = "0.18"
+lalrpop = "0.22.2"
+lalrpop-util = { version = "0.22.2", features = ["lexer"] }
+leon = "3.0.2"
+lettre = { version = "0.11.19", default-features = false, features = [
     "smtp-transport",
     "tokio1-native-tls",
     "tokio1",
     "builder",
 ] }
-libc = "0.2.151"
-libtest-mimic = "0.6.1"
-libsqlite3-sys = { version = "0.27" }
-log = { version = "0.4.20", features = ["std"] }
-nix = { version = "0.27.1", features = [
+libc = "0.2"
+libtest-mimic = "0.8.1"
+libsqlite3-sys = { version = "0.35" }
+log = { version = "0.4", features = ["std"] }
+nix = { version = "0.30.1", features = [
     "fs",
     "process",
     "time",
@@ -160,37 +156,39 @@ nix = { version = "0.27.1", features = [
     "signal",
     "net",
 ] }
-num_cpus = "1.16"
-nut = "0.1.3"
-openssl = "0.10.61"
-proc-macro2 = "1.0.70"
-procfs = { version = "0.16.0", default-features = false }
-quote = "1.0.33"
-rand = { version = "0.8.5" }
-regex = "1.10.2"
-reqwest = { version = "0.12.9", default-features = false, features = [
+num_cpus = "1.17"
+nut = "0.1.4"
+openssl = "0.10.73"
+proc-macro2 = "1.0"
+procfs = { version = "0.18.0", default-features = false }
+quote = "1.0.41"
+rand = { version = "0.9.2" }
+regex = "1.12.2"
+reqwest = { version = "0.12.24", default-features = false, features = [
     "blocking",
     "json",
     "rustls-tls",
 ] }
-rust-ini = "0.20.0"
-semver = { version = "1.0.20", features = ["serde"] }
-serde = { version = "1.0.193", features = ["derive"] }
-serde_json = "1.0.108"
-serde_yaml = "0.9.27"
-signal-hook = "0.3.17"
-strum = { version = "0.25", features = ["derive"] }
-syn = "2.0.41"
-sys-mount = { version = "2.1.0", default-features = false }
+rust-ini = "0.21"
+semver = { version = "1.0", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.145"
+serde_yaml = "0.9"
+signal-hook = "0.3.18"
+strum = { version = "0.27.2", features = ["derive"] }
+syn = "2.0"
+sys-mount = { version = "3.0.1", default-features = false }
 tar = "0.4"
-thiserror = "2.0.9"
-toml_edit = { version = "0.15.0", features = ["easy"] }
-tokio = { version = "1.42", features = ["full"] }
+thiserror = "2"
+tokio = { version = "1.48.0", features = ["full"] }
 tokio-fd = "0.3.0"
-tokio-tungstenite = "0.26.1"
-uuid = { version = "1.6.1", features = ["v4"] }
-which = "5.0.0"
-xshell = "0.2.5"
+tokio-tungstenite = "0.28.0"
+uuid = { version = "1.18.1", features = ["v4"] }
+which = "8.0.0"
+xshell = "0.2.7"
+vergen-gitcl = "1.0.8"
+clap-verbosity-flag = "3.0.4"
+toml = "0.9.8"
 
 [profile.release]
 lto = true
