+++ crates/bpf-filtering/Cargo.toml
@@ -19,7 +19,6 @@ bpf-common = { workspace = true }
 pulsar-core = { workspace = true }
 regex = { workspace = true }
 which = { workspace = true, optional = true }
-lazy_static = { workspace = true }
 
 [build-dependencies]
 bpf-builder = { workspace = true }
