+++ crates/validatron/derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "validatron-derive"
-version = "0.9.0"
+version = "0.10.0-alpha.1"
 license = "Apache-2.0"
 edition = "2021"
