+++ examples/pulsar-extension-module/main.rs
@@ -1,16 +1,20 @@
+use clap::Parser;
+use clap_verbosity_flag::{InfoLevel, Verbosity};
+use pulsar::pulsard::PulsarDaemonOpts;
+
 mod my_custom_module;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Parse cli and handle clap errors
-    let options = pulsar::cli::parse_from_args();
+    let options = Opts::parse();
 
     // Override the default log_level if there is a greater verbosity flag
-    pulsar::init_logger(options.override_log_level);
+    pulsar::init_logger(Some(options.verbosity.log_level_filter()));
 
-    // Run pulsar-exec with the additional module
+    // Run pulsard with the additional module
     #[allow(clippy::blocks_in_conditions)]
-    match pulsar::run_pulsar_exec_custom(&options, |starter| {
+    match pulsar::pulsard::pulsar_daemon_run(&options.daemon_opts, |starter| {
         starter.add_module(my_custom_module::MyCustomModule)?;
 
         Ok(())
@@ -19,8 +23,17 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     {
         Ok(_) => std::process::exit(0),
         Err(e) => {
-            pulsar::cli::report_error(&e);
+            pulsar::utils::report_error(&e);
             std::process::exit(1);
         }
     }
+}
+
+#[derive(Debug, Parser)]
+struct Opts {
+    #[command(flatten)]
+    daemon_opts: PulsarDaemonOpts,
+
+    #[command(flatten)]
+    pub verbosity: Verbosity<InfoLevel>,
 }
