+++ src/utils.rs
@@ -1,0 +1,26 @@
+use std::env;
+
+fn show_backtrace() -> bool {
+    if log::max_level() >= log::LevelFilter::Debug {
+        return true;
+    }
+
+    if let Ok(true) = env::var("RUST_BACKTRACE").map(|s| s == "1") {
+        return true;
+    }
+
+    false
+}
+
+pub fn report_error(e: &anyhow::Error) {
+    // NB: This shows one error: even for multiple causes and backtraces etc,
+    // rather than one per cause, and one for the backtrace. This seems like a
+    // reasonable tradeoff, but if we want to do differently, this is the code
+    // hunk to revisit, that and a similar build.rs auto-detect glue as anyhow
+    // has to detect when backtrace is available.
+    if show_backtrace() {
+        log::error!("{:?}", e);
+    } else {
+        log::error!("{:#}", e);
+    }
+}
