--- vagrant/.gitignore
@@ -1,4 +1,0 @@
-.vagrant
-tests
-pulsar-exec
-test-suite
