--- vagrant/archlinux/Vagrantfile
@@ -1,4 +1,0 @@
-Vagrant.configure("2") do |config|
-  config.vm.box = "archlinux/archlinux"
-  #config.vm.synced_folder ".", "/vagrant", disabled: false
-end
