--- vagrant/ubuntu2010/Vagrantfile
@@ -1,6 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-Vagrant.configure("2") do |config|
-  config.vm.box = "generic/ubuntu2010"
-  config.vm.synced_folder ".", "/vagrant", disabled: false
-end
