--- vagrant/ubuntu2204/Vagrantfile
@@ -1,6 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-Vagrant.configure("2") do |config|
-  config.vm.box = "generic/ubuntu2204"
-  config.vm.synced_folder ".", "/vagrant", disabled: false
-end
